Extract animation component and move loading screen
