+++ src/components/application-loader/application-loader.module.scss
@@ -5,8 +5,6 @@
  */
 
 .loader {
-  @import "./animations.scss";
-
   height: 100%;
   width: 100%;
