+++ src/components/application-loader/loading-screen/animated-hedge-doc-logo/animations.module.scss
@@ -4,18 +4,6 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-@keyframes animation-roll {
-  0% {
-    transform: translateX(calc(-100vw / 2)) rotateZ(0deg);
-  }
-  50% {
-    transform: translateX(calc(100vw / 2)) rotateZ(720deg);
-  }
-  100% {
-    transform: translateX(calc(-100vw / 2)) rotateZ(0deg);
-  }
-}
-
 @keyframes animation-jump {
   0% {
     transform: scale(1, 1) translateY(0);
@@ -74,14 +62,6 @@
   100% {
     transform: translate(1px, -2px) rotate(-1deg);
   }
-}
-
-.animation-roll {
-  transform-origin: center center;
-  animation-duration: 4s;
-  animation-iteration-count: infinite;
-  animation-name: animation-roll;
-  animation-timing-function: linear;
 }
 
 .animation-jump {
